add sorting users with algolia to leaderboards and unfiy tournaments, games and leaderboard pages layouts
