+++ package.json
@@ -36,7 +36,8 @@
     "react-tabs": "^5.1.0",
     "react-toastify": "^9.0.5",
     "styled-components": "^5.2.3",
-    "styled-css-grid": "^1.2.1"
+    "styled-css-grid": "^1.2.1",
+    "unique-names-generator": "^4.7.1"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.16.0",
